+++ src/main.rs
@@ -2,8 +2,10 @@ mod module_examples;
 
 mod primative_examples;
 
-fn main() {
-    module_examples::checknetwork();
 
-    //primativeexamples::example_prim();
+fn main() {
+
+    module_examples::checknetwork();
+    primative_examples::example_prim();
+    
 }