+++ src/primative_examples/arrays_ex.rs
@@ -1,15 +1,57 @@
+use std::mem;
+
 fn array_ex1(){
     let array1 : [i16; 5] = [1,2,4,5,8];
     println!("Array i16 - {}", array1[2]);
 }
 
-pub fn array_exAll(){
+pub fn array_ex_all(){
 array_ex1();
 array_02();
+array_slices_ex01()
 }
 
 fn array_02(){
     let arr: [i32; 5] = [1, 2, 3, 4, 5];
     println!("The value of a is: {}", arr[1]);
 
+}
+
+fn array_slices_ex01(){
+
+    // Fixed-size array (type signature is superfluous)
+    let xs: [i32; 5] = [1, 2, 3, 4, 5];
+
+    // All elements can be initialized to the same value
+    let ys: [i32; 500] = [0; 500];
+
+    // Indexing starts at 0
+    println!("first element of the array: {}", xs[0]);
+    println!("second element of the array: {}", xs[1]);
+
+    // `len` returns the size of the array
+    println!("array size: {}", xs.len());
+
+    // Arrays are stack allocated
+    println!("array occupies {} bytes", mem::size_of_val(&xs));
+
+    // Arrays can be automatically borrowed as slices
+    println!("borrow the whole array as a slice");
+    analyze_slice(&xs);
+
+    // Slices can point to a section of an array
+    // They are of the form [starting_index..ending_index]
+    // starting_index is the first position in the slice
+    // ending_index is one more than the last position in the slice
+    println!("borrow a section of the array as a slice");
+    analyze_slice(&ys[1 .. 4]);
+
+    // Out of bound indexing causes compile error
+    //println!("{}", xs[5]);
+}
+
+// This function borrows a slice
+fn analyze_slice(slice: &[i32]) {
+    println!("first element of the slice: {}", slice[0]);
+    println!("the slice has {} elements", slice.len());
 }